fix amount to float

+++ src/models/stock_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import CheckConstraint, Column, Index, Integer, String
+from sqlalchemy import CheckConstraint, Column, Index, Integer, String, Numeric
 
 from src.config.database import Base
 
@@ -7,7 +7,7 @@ class StockModel(Base):
     __tablename__ = "stocks"
 
     company_code = Column(String(10), nullable=False, primary_key=True)
-    purchased_amount = Column(Integer, nullable=False)
+    purchased_amount = Column(Numeric, nullable=False)
 
     __table_args__ = (
         CheckConstraint(
